moved logger to pkg
